Use `hairpin_snat_ip` for Loadbalancers
Currently when traffic is sent to an OVN Lb and is hairpinned
back to the same dstIP as the srcIP the OVN lb randomly chooses
and IP for the SNAT.  This makes it hard for the network policy
code to accurately allow traffic in some scenarios.

This Commit forces the IP chosed for SNAT to be 169.254.169.3 for
ipv4 and fd69::3 for ipv6.  Which will allow us to accurately
allow traffic with network policy without creating a dependency
between the network policy and service objects within OVN-K8s.

The new Loadbalancer is applied to every OVN Loadbalancer to handle
the hairpin traffic + network policy scenario to be handled correctly

Signed-off-by: astoycos <astoycos@redhat.com>