Remove nbctl daemon
Remove the ability/config to run nbctl in daemon mode
in ovn-kubernetes.  This is due to the fact that 90%
of nbctl calls have been converted to using libovsdb clients
in the ovnkube-master process rendering the daemon obsolete.

Fix the GH actions workflow to ensure we use the MASTER kind
scripts and images when bringing up the pre-upgrade kind cluster

And checkout the PR's code to ensure we build/deploy updated manifests
as well as images.

Signed-off-by: astoycos <astoycos@redhat.com>